--- packages/x-components/src/types/result-location.ts
@@ -1,13 +1,0 @@
-/**
- * Indicates where the result where the event comes from is located.
- *
- * @public
- */
-export type ResultLocation =
-  | 'recommendations'
-  | 'predictive_layer'
-  | 'no_results'
-  | 'results'
-  | 'next_query'
-  | 'identifier_result'
-  | 'discovery_wall';
