+++ packages/x-components/src/x-modules/search/store/types.ts
@@ -12,7 +12,7 @@ import {
 } from '@empathyco/x-types';
 import { XActionContext, XStoreModule } from '../../../store';
 import { StatusMutations, StatusState } from '../../../store/utils/status-store.utils';
-import { QueryOrigin } from '../../../types/query-origin';
+import { Origin } from '../../../types/origin';
 import { UrlParams } from '../../../types/url-params';
 import { Dictionary } from '../../../utils/types';
 import { SearchConfig } from '../config.types';
@@ -32,7 +32,7 @@ export interface SearchState extends StatusState {
   /** A flag to indicate if new results are append to the current instead of replacing them. */
   isAppendResults: boolean;
   /** The origin property of the request. */
-  origin: QueryOrigin | null;
+  origin: Origin | null;
   /** The current page of the request. */
   page: number;
   /** The extra params property of the state. */
@@ -106,7 +106,7 @@ export interface SearchMutations extends StatusMutations {
    *
    * @param origin - The new origin.
    */
-  setOrigin(origin: QueryOrigin | undefined | null): void;
+  setOrigin(origin: Origin | undefined | null): void;
   /**
    * Sets the page of the module.
    *
